Split vpype into two separate packages.
`vpype`: contains the core of vpype
`vpype_cli`: contains the click cli and all the commands
