+++ setup.py
@@ -1,4 +1,4 @@
-from setuptools import setup, find_packages
+from setuptools import setup
 
 
 with open("README.md") as f:
@@ -16,7 +16,7 @@ setup(
     author="Antoine Beyeler",
     url="https://github.com/abey79/vpype",
     license=license,
-    packages=find_packages(exclude=("examples", "tests")),
+    packages=["vpype", "vpype_cli"],
     install_requires=[
         'Click',
         'click-plugins',
@@ -28,6 +28,6 @@ setup(
     ],
     entry_points='''
         [console_scripts]
-        vpype=vpype:cli
+        vpype=vpype_cli.cli:cli
     '''
 )
