+++ vpype/__init__.py
@@ -1,18 +1,4 @@
-from .vpype import cli
-from .decorators import layer_processor, global_processor, block_processor, generator
-from .model import LineCollection
-
-# register all commands
-from .blocks import *
-from .debug import *
-from .frames import *
-from .generators import *
-from .operations import *
-from .primitives import *
-from .read import *
-from .script import *
-from .show import *
-from .transforms import *
-from .write import *
-
-__all__ = ["cli", "layer_processor", "global_processor", "generator", "LineCollection"]
+from .decorators import *
+from .layers import *
+from .model import *
+from .utils import *
