--- vpype/__main__.py
@@ -1,4 +1,0 @@
-# from .vpype import cli
-import vpype
-
-vpype.cli()
