+++ vpype_cli/__init__.py
@@ -1,0 +1,12 @@
+# register all commands
+from .blocks import *
+from .debug import *
+from .frames import *
+from .generators import *
+from .operations import *
+from .primitives import *
+from .read import *
+from .script import *
+from .show import *
+from .transforms import *
+from .write import *
